code/requirements.txt
gunicorn
gevent
flask
jieba
wordcloud
numpy
matplotlib
Pillow
